Move some test data to __tests__/data.ts
