+++ src/domain/checklist/__tests__/uploadToS3Checklist.test.ts
@@ -1,28 +1,6 @@
-import * as O from 'fp-ts/Option';
-import { PreLoadRecord } from '../../PreLoadRepository';
-import { UploadToS3Record } from '../../UploadToS3RecordRepository';
 import { evalCheck } from '../types';
 import { check0, check1 } from '../uploadToS3Checklist';
-
-const preLoadResponse = { preloadIdx: '0', secret: 'a-secret', url: 'a-url', key: 'a-key' };
-const preLoadBody = { preloadIdx: '0', contentType: 'application/pdf', sha256: 'a-sha256' };
-
-const preLoadRecord: PreLoadRecord = {
-  type: 'PreLoadRecord',
-  input: { apiKey: 'an-api-key', body: [preLoadBody] },
-  output: { statusCode: 200, returned: [preLoadResponse] },
-};
-
-const baseUploadToS3Record: UploadToS3Record = {
-  type: 'UploadToS3Record',
-  input: {
-    key: preLoadResponse.key,
-    checksumAlg: O.none,
-    secret: preLoadResponse.secret,
-    checksum: preLoadBody.sha256,
-  },
-  output: { statusCode: 200, returned: 10 },
-};
+import * as data from '../../__tests__/data';
 
 const group = { name: 'Group' };
 
@@ -30,20 +8,20 @@ describe('uploadToS3Checklist', () => {
   it('should exist a response with status code 2xx', () => {
     const check = evalCheck({ ...check0, group });
 
-    const actualOK = check([baseUploadToS3Record]);
+    const actualOK = check([data.uploadToS3Record]);
     expect(actualOK.result).toStrictEqual('ok');
 
-    const actualKO = check([preLoadRecord]);
+    const actualKO = check([data.preLoadRecord]);
     expect(actualKO.result).toStrictEqual('ko');
   });
 
   it('should exist a correlation between the "preload"', () => {
     const check = evalCheck({ ...check1, group });
 
-    const actualOK = check([preLoadRecord, baseUploadToS3Record]);
+    const actualOK = check([data.preLoadRecord, data.uploadToS3Record]);
     expect(actualOK.result).toStrictEqual('ok');
 
-    const actualKO = check([baseUploadToS3Record, baseUploadToS3Record]);
+    const actualKO = check([data.uploadToS3Record, data.uploadToS3Record]);
     expect(actualKO.result).toStrictEqual('ko');
   });
 });
